+++ src/admin/messages/fr/newsadmin.php
@@ -33,4 +33,10 @@ return [
 // 3.1.0
 
     'preview' => 'Preview',
+
+// 4.0.0
+
+    'article_author' => 'Author',
+    'article_link' => 'Link',
+    'article_is_archived' => 'Archived',
 ];