+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 In order to read more about upgrading and BC breaks have a look at the [UPGRADE Document](UPGRADE.md).
 
+## 4.0.0 (7. July 2021)
+
++ [#24](https://github.com/luyadev/luya-module-news/pull/24) New field to archive certain News Articles. A new archive action is available where developers can list the archived entries.
+
 ## 3.1.0 (14. June 2021)
 
-+ [#23](https://github.com/luyadev/luya-module-news/pull/23) Preview Active Window for News
++ [#23](https://github.com/luyadev/luya-module-news/pull/23) Preview Active Window for News Articles.
 
 ## 3.0.0 (7. April 2021)